Add a global lock.
Sometimes the sqlite3 database connection started to hang, when more
than one process tried to delete the same set of rows. Also, this should
help avoid problems when multiple threads use the Python API: plugins
are initialized just at the moment they are going to be used which could
lead to problems when two threads try to setup the same plugin at the
same time.